clock-api/parse-lineup/requirements.txt
google-genai==1.0.0
firebase-admin==6.5.0
pydantic==2.9.2
boto3==1.35.0
